requirements.txt
Django==1.4
Fabric==1.4.1
South==0.7.4
Sphinx==1.1.3
boto==2.3.0
django-celery==2.5.5
django-compressor==1.1.2
django-debug-toolbar==0.9.4
django-htmlmin==0.5.1
django-kombu==0.9.4
django-pylibmc-sasl==0.2.4
django-storages==1.1.4
gunicorn==0.14.2
newrelic==1.2.1.265
psycopg2==2.4.2
pylibmc==1.2.3
raven==1.7.5
django-model-utils
PIL
lxml
sorl-thumbnail
django-friendship

-e git+git@github.com:toastdriven/django-tastypie.git@perms#egg=django_tastypie-dev
-e git+https://github.com/GraylinKim/sc2reader.git@6b540c27999dd06f06dbd16550ea5a0a54fdedc1#egg=sc2reader
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django_registration
